requirements.txt
adal==1.2.7
alembic==1.14.0
argcomplete==3.5.2
attrs==24.3.0
azure-ai-ml==1.23.0
azure-common==1.1.28
azure-core==1.32.0
azure-graphrbac==0.61.2
azure-identity==1.19.0
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.5.0
azure-mgmt-keyvault==10.3.1
azure-mgmt-network==28.0.0
azure-mgmt-resource==23.2.0
azure-mgmt-storage==21.2.1
azure-storage-blob==12.19.0
azure-storage-file-datalake==12.14.0
azure-storage-file-share==12.20.0
azureml-core==1.59.0
azureml-mlflow==1.58.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
databricks-sdk==0.39.0
Deprecated==1.2.15
docker==7.1.0
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.3
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-auth==2.37.0
googleapis-common-protos==1.66.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
gunicorn==23.0.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
isodate==0.7.2
itsdangerous==2.2.0
jeepney==0.8.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpickle==4.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
knack==0.12.0
Mako==1.3.8
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.10.0
mlflow==2.19.0
mlflow-skinny==2.19.0
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
msrestazure==0.6.4.post1
ndg-httpsclient==0.5.1
networkx==3.4.2
numpy==2.2.0
oauthlib==3.2.2
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
pandas==2.2.3
paramiko==3.5.0
pathspec==0.12.1
pillow==11.0.0
pkginfo==1.12.0
portalocker==2.10.1
proto-plus==1.25.0
protobuf==5.29.1
psutil==6.1.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydash==8.0.4
Pygments==2.18.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==24.3.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.6.0
scipy==1.14.1
SecretStorage==3.3.3
six==1.17.0
smmap==5.0.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
strictyaml==1.7.3
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
torch==2.5.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
zipp==3.21.0
